requirements.txt
Pillow
pytesseract
pandas
openpyxl
requests
TikTokApi
instaloader
fastapi
uvicorn[standard]
python-multipart